requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
fastapi-mail==1.4.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1